+++ packages/metadata-builders/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Add `viewFns` checksum and dynamic builder capabilities.
+
 ### Changed
 
 - `getLookupFn` takes a `UnifiedMetadata`, instead of `V14 | V15`.